# conf: cache getMode