--- a/internal/conf/conf.go
+++ b/internal/conf/conf.go
@@ -54,7 +54,19 @@ const (
 	modeEmpty
 )
 
+var (
+	cachedModeOnce sync.Once
+	cachedMode     configurationMode
+)
+
 func getMode() configurationMode {
+	cachedModeOnce.Do(func() {
+		cachedMode = getModeUncached()
+	})
+	return cachedMode
+}
+
+func getModeUncached() configurationMode {
 	mode := os.Getenv("CONFIGURATION_MODE")
 
 	switch mode {