@keegancsmith 2022-09-05
> Thanks for the PR description, it never clicked with me to consider why we panic on `internal/env.Get` post-initialization! LGTM

That decision existed because we wanted to enforce all envvars we read are known at startup so that we can correctly produce a help output of all config envvars (and validate). Unfortunetly this is context that exists from conversations I had, found the original PR and it has zero description lol https://github.com/sourcegraph/enterprise/pull/2365